[expo-cli] add progress bar for project upload
